scheduler/requirements.txt
APScheduler==3.8.1
psycopg2==2.9.1
python-decouple==3.6
elasticsearch==7.17.0
pymongo==4.0.1
backoff==1.11.1
pydantic==1.9.0